linuxcnc: create unified probe routine
This routine can probe the X, Y, or Z axis based on subroute parameter.
This avoids duplicating the entire probing logic into 3 separate files.

This subroutine is triggered from the existing 3 homing routines, to
avoid having to edit the LinuxCNC HAL file to change the MDI commands in
this commit. This also allows easily reverting the change if necessary.